Gabime/release 0.33.3 (#154)
Update to version 0.33.3

+++ src/FwUpdate/Cmds.cc
@@ -14,7 +14,7 @@ std::string dlver()
     return "\ndlver";
 }
 
-std::string dlspd(uint32_t spd)
+std::string dlspd(long spd)
 {
     std::ostringstream oss;
     oss << "\ndlspd " << spd;
